backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
gitpython>=3.1.32
google-cloud-storage>=2.10.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
